Finished project of 14 sprint (tests: 1)

+++ controllers/card.js
@@ -1,73 +1,77 @@
 const Card = require('../models/card');
-const { ERROR_SERV, ERROR_VALID, ERROR_NOT_FOUND } = require('../utils/constants');
+const ValidationError = require('../utils/errors/ValidationError');
+const NotFoundError = require('../utils/errors/NotFoundError');
+const СredentialError = require('../utils/errors/СredentialError');
 
-function getCards(req, res) {
+function getCards(req, res, next) {
   Card.find({})
     .then((cards) => res.send({ cards }))
-    .catch(() => res.status(ERROR_SERV).send({ message: 'Произошла ошибка на сервере' }));
+    .catch(next);
 }
 
-function createCard(req, res) {
+function createCard(req, res, next) {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send({ card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(ERROR_VALID).send({ message: 'Переданны некорректные данные при создании карточки' });
-      } else {
-        res.status(ERROR_SERV).send({ message: 'Произошла ошибка на сервере' });
+        throw new ValidationError('Переданны некорректные данные при создании карточки');
       }
+      return next(err);
     });
 }
 
-function deleteCard(req, res) {
-  Card.findByIdAndRemove(req.params.cardId)
+function deleteCard(req, res, next) {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        return res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка не найдена' });
+        throw new NotFoundError('Запрашиваемая карточка не найдена');
       }
-      return res.send({ card });
-    })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(ERROR_VALID).send({ message: 'Переданны некорректные данные при удалении карточки' });
+      if (card.owner.toString() === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => res.send({ card }))
+          .catch((err) => {
+            if (err.name === 'CastError') {
+              throw new ValidationError('Переданны некорректные данные при удалении карточки');
+            }
+            return next(err);
+          });
       } else {
-        res.status(ERROR_SERV).send({ message: 'Произошла ошибка на сервере' });
+        throw new СredentialError('Недостаточно прав');
       }
-    });
+    })
+    .catch(next);
 }
 
-function putLikeCard(req, res) {
+function putLikeCard(req, res, next) {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        return res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка не найдена' });
+        throw new NotFoundError('Запрашиваемая карточка не найдена');
       }
       return res.send({ card });
     })
     .catch((err) => {
       if (err.name === 'ValidationError' || err.name === 'CastError') {
-        res.status(ERROR_VALID).send({ message: 'Переданны некорректные данные при добавлении лайка у карточки' });
-      } else {
-        res.status(ERROR_SERV).send({ message: 'Произошла ошибка на сервере' });
+        throw new ValidationError('Переданны некорректные данные при добавлении лайка у карточки');
       }
+      return next(err);
     });
 }
 
-function deleteLikeCard(req, res) {
+function deleteLikeCard(req, res, next) {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        return res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка не найдена' });
+        throw new NotFoundError('Запрашиваемая карточка не найдена');
       }
       return res.send({ card });
     })
     .catch((err) => {
       if (err.name === 'ValidationError' || err.name === 'CastError') {
-        res.status(ERROR_VALID).send({ message: 'Переданны некорректные данные при удалении лайка у карточки' });
-      } else {
-        res.status(ERROR_SERV).send({ message: 'Произошла ошибка на сервере' });
+        throw new ValidationError('Переданны некорректные данные при добавлении лайка у карточки');
       }
+      return next(err);
     });
 }
